requirements.txt
absl-py==2.1.0
asttokens==3.0.0
av==13.1.0
beautifulsoup4==4.12.3
chex==0.1.88
click==8.1.8
cloup==3.0.5
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
etils==1.11.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
glcontext==3.0.0
ipykernel==6.29.5
ipython==8.30.0
isosurfaces==0.1.2
jax==0.4.38
jax-cuda12-pjrt==0.4.38
jax-cuda12-plugin==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.43.0
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optax==0.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.27.0
pydub==0.25.1
pyglet==2.1.1
PyGLM==2.7.3
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
screeninfo==0.8.1
setuptools==75.6.0
six==1.17.0
skia-pathops==0.8.0.post2
soupsieve==2.6
srt==3.5.3
stack-data==0.6.3
svgelements==1.9.6
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
watchdog==6.0.0
wcwidth==0.2.13